requirements.txt
tensorflow>=2.11.0
opencv-python>=4.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.3.0
tqdm>=4.62.0
Pillow>=8.0.0
mediapipe>=0.10.0
flask>=2.0.0
flask-cors>=3.0.0 